+++ endpoint.js
@@ -31,6 +31,7 @@ module.exports = {
   SuffixArray: SuffixArray,
   GeneralizedSuffixArray: SuffixArray.GeneralizedSuffixArray,
   Set: require('./set.js'),
+  SparseSet: require('./sparse-set.js'),
   SymSpell: require('./symspell.js'),
   Trie: require('./trie.js'),
   VPTree: require('./vp-tree.js')