+++ sparse-set.js
@@ -1,0 +1,162 @@
+/**
+ * Mnemonist SparseSet
+ * ====================
+ *
+ * JavaScript sparse set implemented on top of a Uint32Array.
+ *
+ * [Reference]: https://research.swtch.com/sparse
+ */
+
+/**
+ * SparseSet.
+ *
+ * @constructor
+ */
+function SparseSet(length) {
+
+  // Properties
+  this.size = 0;
+  this.length = length;
+  this.dense = new Uint32Array(length);
+  this.sparse = new Uint32Array(length);
+}
+
+/**
+ * Method used to clear the structure.
+ *
+ * @return {undefined}
+ */
+SparseSet.prototype.clear = function() {
+  this.size = 0;
+};
+
+/**
+ * Method used to check the existence of a member in the set.
+ *
+ * @param  {number} member - Member to test.
+ * @return {SparseSet}
+ */
+SparseSet.prototype.has = function(member) {
+  var index = this.sparse[member];
+
+  return (
+    index < this.size &&
+    this.dense[index] === member
+  );
+};
+
+/**
+ * Method used to add a member to the set.
+ *
+ * @param  {number} member - Member to add.
+ * @return {SparseSet}
+ */
+SparseSet.prototype.add = function(member) {
+  if (this.has(member))
+    return this;
+
+  this.dense[this.size] = member;
+  this.sparse[member] = this.size;
+  this.size++;
+
+  return this;
+};
+
+/**
+ * Method used to remove a member from the set.
+ *
+ * @param  {number} member - Member to delete.
+ * @return {SparseSet}
+ */
+SparseSet.prototype.delete = function(member) {
+  if (!this.has(member))
+    return;
+
+  var index = this.dense[this.size - 1];
+  this.dense[this.sparse[member]] = index;
+  this.sparse[index] = this.sparse[member];
+  this.size--;
+
+  return this;
+};
+
+/**
+ * Method used to iterate over the set's values.
+ *
+ * @param  {function}  callback - Function to call for each item.
+ * @param  {object}    scope    - Optional scope.
+ * @return {undefined}
+ */
+SparseSet.prototype.forEach = function(callback, scope) {
+  scope = arguments.length > 1 ? scope : this;
+
+  var item;
+
+  for (var i = 0; i < this.size; i++) {
+    item = this.dense[i];
+
+    callback.call(scope, item, item);
+  }
+};
+
+/**
+ * Sparse Set Iterator class.
+ */
+function SparseSetIterator(next) {
+  this.next = next;
+}
+
+/**
+ * Method used to create an iterator over a set's values.
+ *
+ * @return {Iterator}
+ */
+SparseSet.prototype.values = function() {
+  var size = this.size,
+      dense = this.dense,
+      i = 0;
+
+  return new SparseSetIterator(function() {
+    if (i < size) {
+      var item = dense[i];
+      i++;
+
+      return {
+        value: item
+      };
+    }
+
+    return {
+      done: true
+    };
+  });
+};
+
+/**
+ * Attaching the #.values method to Symbol.iterator if possible.
+ */
+if (typeof Symbol !== 'undefined')
+  SparseSet.prototype[Symbol.iterator] = SparseSet.prototype.values;
+
+/**
+ * Convenience known methods.
+ */
+SparseSet.prototype.inspect = function() {
+  var proxy = new Set();
+
+  for (var i = 0; i < this.size; i++)
+    proxy.add(this.dense[i]);
+
+  // Trick so that node displays the name of the constructor
+  Object.defineProperty(proxy, 'constructor', {
+    value: SparseSet,
+    enumerable: false
+  });
+
+  return proxy;
+};
+
+/**
+ * Exporting.
+ */
+module.exports = SparseSet;