+++ test/sparse-set.js
@@ -1,0 +1,76 @@
+/**
+ * Mnemonist SparseSet Unit Tests
+ * ===============================
+ */
+var assert = require('assert'),
+    SparseSet = require('../sparse-set.js'),
+    obliterator = require('obliterator');
+
+describe('SparseSet', function() {
+
+  it('should be possible to add items to the set.', function() {
+    var set = new SparseSet(10);
+
+    set.add(3);
+    set.add(4);
+    set.add(3);
+
+    assert.strictEqual(set.size, 2);
+    assert.strictEqual(set.length, 10);
+  });
+
+  it('should be possible to test the existence of items in the set.', function() {
+    var set = new SparseSet(10);
+
+    set.add(3);
+    assert.strictEqual(set.has(3), true);
+    assert.strictEqual(set.has(1), false);
+  });
+
+  it('should be possible to delete items from the set.', function() {
+    var set = new SparseSet(10);
+
+    set.add(3);
+    set.delete(3);
+    set.delete(4);
+
+    assert.strictEqual(set.size, 0);
+  });
+
+  it('should be possible to clear the set.', function() {
+    var set = new SparseSet(10);
+
+    for (var i = 0; i < 6; i++)
+      set.add(i);
+
+    set.clear();
+
+    assert.strictEqual(set.size, 0);
+    assert.strictEqual(set.has(1), false);
+  });
+
+  it('should be possible to iterate over the set\'s items.', function() {
+    var set = new SparseSet(10);
+
+    set.add(3);
+    set.add(6);
+    set.add(9);
+
+    var array = [3, 6, 9],
+        i = 0;
+
+    set.forEach(function(number) {
+      assert.strictEqual(number, array[i++]);
+    });
+  });
+
+  it('should be possible to create an iterator over the set\'s values.', function() {
+    var set = new SparseSet(10);
+
+    set.add(3);
+    set.add(6);
+    set.add(9);
+
+    assert.deepEqual(obliterator.consume(set.values()), [3, 6, 9]);
+  });
+});