+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## O.14.0 (provisional)
 
+* Adding `SparseSet`.
 * Adding iterator methods to `BitSet`.
 
 ## 0.13.0